Use same lock file for whole app & move appsignal to healthlocker
Appsignal wasn't available in healthlocker/endpoint.ex because it's only at the top level of the project. Additionally the setup in apps/healthlocker/mix.exs meant that you had to get the deps separately when cd'd into there and had a separate lock file. Updating the setup means that the deps and lock file are all in one place, and mix deps.get only has to be run once from the root of the umbrella app.